Fix CMakeLists files
It's probably easier and cleaner to do some globbing to get the list of
header and source files "from disk" instead of listing them by hand.

Here's a proposed fix of the CMakeLists file. It will generate a working
makefile on linux, automatically running Qt's MOC  and should work the
same way on Windows.

Signed-off-by: Arthur Brainville (Ybalrid) <ybalrid@ybalrid.info>